requirements.txt
kafka-python==2.2.11
mysql-connector-python
requests